+++ middlewares/validatId.js
@@ -1,4 +1,5 @@
 const { isValidObjectId } = require("mongoose");
+const Joi = require('joi');
 
 const validatIdParam = (req, res, next) => {
     const { contactId } = req.params;
@@ -9,4 +10,28 @@ const validatIdParam = (req, res, next) => {
     next();
 };
 
-module.exports = validatIdParam;
+const joiRegisterSchema = Joi.object({
+    name: Joi.string().required(),
+    email: Joi.string().required(),
+    password: Joi.string().min(6).required()
+});
+
+const joiLoginSchema = Joi.object({
+    email: Joi.string().required(),
+    password: Joi.string().min(6).required()
+});
+
+const joiUpdSubscriptionSchema = Joi.object({
+    subscription: Joi.string().valid("starter", "pro", "business").required()
+});
+
+const schemas = {
+    joiRegisterSchema,
+    joiLoginSchema,
+    joiUpdSubscriptionSchema
+};
+
+module.exports = {
+    validatIdParam,
+    schemas
+};
