+++ models/user.js
@@ -1,5 +1,5 @@
 const { Schema, model } = require('mongoose');
-const Joi = require('joi');
+// const Joi = require('joi');
 
 const emailRegexp = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 
@@ -26,29 +26,6 @@ const userSchema = new Schema({
     },
 }, { versionKey: false, timestamps: true });
 
-const joiRegisterSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    password: Joi.string().min(6).required()
-});
-
-const joiLoginSchema = Joi.object({
-    email: Joi.string().required(),
-    password: Joi.string().min(6).required()
-});
-
-const joiUpdSubscriptionSchema = Joi.object({
-    subscription: Joi.string().valid("starter", "pro", "business").required()
-});
-
-const schemas = {
-    joiRegisterSchema,
-    joiLoginSchema,
-    joiUpdSubscriptionSchema
-};
 const User = model("user", userSchema);
 
-module.exports = {
-    User,
-    schemas
-};
+module.exports = User;
